cli: move commands to cli/cmd
