+++ cli/cmd/check_config.go
@@ -1,4 +1,4 @@
-package cli
+package cmd
 
 import (
 	"fmt"
