+++ cli/cmd/check_config_test.go
@@ -1,4 +1,4 @@
-package cli
+package cmd
 
 import (
 	"testing"
