+++ cli/cmd/root.go
@@ -1,4 +1,4 @@
-package cli
+package cmd
 
 import (
 	"io/ioutil"
