+++ cli/cmd/silence_add.go
@@ -1,4 +1,4 @@
-package cli
+package cmd
 
 import (
 	"bytes"
