+++ cli/cmd/silence_import.go
@@ -1,4 +1,4 @@
-package cli
+package cmd
 
 import (
 	"encoding/json"
