+++ cli/cmd/silence_query.go
@@ -1,4 +1,4 @@
-package cli
+package cmd
 
 import (
 	"encoding/json"
