+++ cli/cmd/testdata/conf.bad.yml
