+++ cmd/amtool/main.go
@@ -1,7 +1,7 @@
 package main
 
-import "github.com/prometheus/alertmanager/cli"
+import "github.com/prometheus/alertmanager/cli/cmd"
 
 func main() {
-	cli.Execute()
+	cmd.Execute()
 }
